==> notebooks/3.0-syed-modelling.ipynb <==
import torch
import gc
import pandas as pd
from torch.utils.data import TensorDataset, DataLoader
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM

data = pd.read_csv('../data/processed/lyrics_processed.csv')

train_data = []
for i in range(len(data)):
    data_elem = data.iloc[i]
    song = data_elem['song']
    artist = data_elem['artist']
    year = data_elem['year']
    genre = data_elem['genre']
    lyrics = data_elem['lyrics']
    
    train_data.append({"input_text": f"Give me the lyrics of a song that was made by {artist} and was releasd in {year} and is of the {genre} genre.", "label_text": lyrics})

train_data, test_data = train_data[:int(0.9*len(train_data))], train_data[int(0.9*len(train_data)):]

tokenizer = AutoTokenizer.from_pretrained("google/flan-t5-large")
model = AutoModelForSeq2SeqLM.from_pretrained("google/flan-t5-large")

tokenized_data = tokenizer([d["input_text"] for d in train_data], padding=True, truncation=True, return_tensors='pt')
input_ids = tokenized_data.input_ids
attention_mask = tokenized_data.attention_mask

label_ids = tokenizer([d["label_text"] for d in train_data], padding=True, truncation=True, return_tensors='pt').input_ids

batch_size = 8

dataset = TensorDataset(input_ids, attention_mask, label_ids)
dataloader = DataLoader(dataset, batch_size=batch_size, shuffle=True)

def train(model, dataloader, optimizer, device, epoch):
    torch.cuda.empty_cache()
    gc.collect()
    model.train()
    total_loss = 0.0
    for idx, batch in enumerate(dataloader):
        input_ids, attention_mask, labels = batch
        input_ids = input_ids.to(device)
        attention_mask = attention_mask.to(device)
        labels = labels.to(device)
        
        optimizer.zero_grad()
        outputs = model(input_ids=input_ids, attention_mask=attention_mask, labels=labels)
        loss = outputs.loss
        loss.backward()
        optimizer.step()

        total_loss += loss.item()
        print(f"Epoch {epoch+1} - Batch {idx*batch_size}/{len(dataloader)} - Loss: {loss.item()}")
        
        model.save_pretrained('../models/flan_t5_large')
        
        del input_ids
        del attention_mask
        del labels
        torch.cuda.empty_cache()
        gc.collect()
    del dataloader
    torch.cuda.empty_cache()
    gc.collect()
        
    avg_loss = total_loss / len(dataloader)
    print(f"Epoch {epoch+1} Complete! - Average Loss: {avg_loss}")

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
model.to(device)

learning_rate = 1e-4
lr_scheduler_step = 1
lr_scheduler_gamma = 0.9
num_epochs = 10

optimizer = torch.optim.AdamW(model.parameters(), lr=learning_rate)
scheduler = torch.optim.lr_scheduler.StepLR(optimizer, step_size=lr_scheduler_step, gamma=lr_scheduler_gamma)

torch.cuda.empty_cache()
gc.collect()
for epoch in range(num_epochs):
    train(model, dataloader, optimizer, device, epoch)
    scheduler.step()
    model.save_pretrained('../models/flan_t5_large')
    torch.cuda.empty_cache()
    gc.collect()
model.save_pretrained('../models/flan_t5_large')
torch.cuda.empty_cache()
gc.collect()
