==> notebooks/4.0-syed-model_testing.ipynb <==
from transformers import AutoTokenizer, AutoModelForSeq2SeqLM

lyrics_tokenizer = AutoTokenizer.from_pretrained("google/flan-t5-large")
lyrics_model = AutoModelForSeq2SeqLM.from_pretrained("../models/flan_t5_large")

# song_name_tokenizer = AutoTokenizer.from_pretrained()
# song_name_model = 

def generate_lyrics(artist, year, genre):
    text = f"Give me the lyrics of a song that was made by {artist} and was releasd in {year} and is of the {genre} genre."
    input_ids = lyrics_tokenizer.encode(text, return_tensors="pt")
    outputs = lyrics_model.generate(input_ids, min_length=75, max_length=100, temperature=0.5, num_beams=5, no_repeat_ngram_size=2, do_sample=True, top_k=10, top_p=0.9, remove_invalid_values=True)
    return lyrics_tokenizer.decode(outputs[0])

# def generate_name_from_lyrics(lyrics):
#     text = f"Give me the name of the song that goes like this: {lyrics}"
#     encoded_text = tokenizer(text, return_tensors="pt")
#     input_ids, attention_mask = encoded_text['input_ids'], encoded_text['attention_mask']
#     outputs = model.generate(input_ids=input_ids, attention_mask=attention_mask, min_length=10, max_length=20, temperature=0.5, num_beams=5, no_repeat_ngram_size=2, do_sample=True, top_k=10, top_p=0.9, remove_invalid_values=True)
#     return tokenizer.decode(outputs[0])

generate_lyrics("Drake", 2010, "Rock")
